abstracting the druid mana usage card
